Split ExecutionContext class into distinct modules
1. the context itself (that may have many threads);
2. the fiber scheduling (on _one_ thread).

The context itself is meant to create and maintain (or monitor) one or
many threads (that each need a scheduler) and the object itself can be
used for cross context communication, for example context A telling
context B to spawn a fiber. Their usage is both external (public API)
and internal (private API).

The schedulers are meant to do the actual scheduling of fibers inside a
single thread. Their usage is mostly internal (private API).

The choice of modules is to be more versatile in the way we want to
implement each execution context. We might want to implement both
modules in a single type (e.g. single threaded context) or as distinct
types.