Convert metadata into YAML frontmatter (#13)
Instead of a unspecific markdown list we can use YAML frontmatter to declare metadata for the RFC document. This is a more formalized data format and makes it easier to extract information from the document. It also renders quite nicely in GitHub.
In the process I'm also simplifying the PR and issue references to keep them cleen URLs (they're rendered as short reference on GitHub).

After merging this we need to patch pending PRs as well to ensure a consistent format.